@@ -1,10 +1,10 @@
 module golang.design/x/tgstore
 
-go 1.13
+go 1.16
 
 require (
 	github.com/allegro/bigcache/v2 v2.2.5 // indirect
 	github.com/allegro/bigcache/v3 v3.0.0
-	golang.org/x/crypto v0.0.0-20201208171446-5f87f3452ae9
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad
 	gopkg.in/tucnak/telebot.v2 v2.3.5
 )
